put tilt series table in own subfolder
