+++ frontend/packages/data-portal/app/components/TiltSeriesTable/index.ts
@@ -1,0 +1,2 @@
+export { TiltSeriesKeys } from './constants'
+export { TiltSeriesTable } from './TiltSeriesTable'
