+++ frontend/packages/data-portal/app/components/TiltSeriesTable/utils.ts
